test: update test cases
