+++ data-providers/jdbc-data-provider/src/main/java/datart/data/provider/jdbc/adapters/JdbcDataProviderAdapter.java
@@ -157,7 +157,7 @@ public class JdbcDataProviderAdapter implements Closeable {
 
     public String getQueryKey(QueryScript script, ExecuteParam executeParam) throws SqlParseException {
         SqlScriptRender render = new SqlScriptRender(script, executeParam, getSqlDialect(), jdbcProperties.isEnableSpecialSql());
-        return "Q" + DigestUtils.md5Hex(render.render(true, true, true));
+        return "Q" + DigestUtils.md5Hex(render.render(true, supportPaging(), true));
     }
 
     protected Column readTableColumn(ResultSet columnMetadata) throws SQLException {
