+++ data-providers/jdbc-data-provider/src/test/java/datart/data/provider/sql/examples/ExecParamSqlExamples.java
@@ -1,9 +1,9 @@
 package datart.data.provider.sql.examples;
 
 import datart.core.data.provider.ExecuteParam;
-import datart.data.provider.sql.entity.SqlTestEntity;
 import datart.data.provider.sql.common.ParamFactory;
 import datart.data.provider.sql.common.TestSqlDialects;
+import datart.data.provider.sql.entity.SqlTestEntity;
 import org.junit.platform.commons.util.StringUtils;
 
 import java.util.ArrayList;
@@ -44,9 +44,9 @@ public class ExecParamSqlExamples {
     }
 
     private static void initExecTemplateMap() {
-        execParamTemplate.put(TestSqlDialects.MYSQL.getDatabaseProduct().name(), "SELECT `DATART_VTABLE`.`name` AS `name`, `DATART_VTABLE`.`age` AS `age`, `DATART_VTABLE`.`id`, SUM(`DATART_VTABLE`.`val`) AS `SUM(val)`  " +
+        execParamTemplate.put(TestSqlDialects.MYSQL.getDatabaseProduct().name(), "SELECT `DATART_VTABLE`.`name` AS `name`, `DATART_VTABLE`.`age` AS `age`, `DATART_VTABLE`.`id` AS `id`, SUM(`DATART_VTABLE`.`val`) AS `SUM(val)`  " +
                 "FROM  ( &#%xxx%#& )  AS `DATART_VTABLE`  GROUP BY `DATART_VTABLE`.`id`  ORDER BY COUNT(`DATART_VTABLE`.`age`) DESC");
-        execParamTemplate.put(TestSqlDialects.ORACLE.getDatabaseProduct().name(), "SELECT \"DATART_VTABLE\".\"name\" \"name\", \"DATART_VTABLE\".\"age\" \"age\", \"DATART_VTABLE\".\"id\", SUM(\"DATART_VTABLE\".\"val\") \"SUM(val)\"  " +
+        execParamTemplate.put(TestSqlDialects.ORACLE.getDatabaseProduct().name(), "SELECT \"DATART_VTABLE\".\"name\" \"name\", \"DATART_VTABLE\".\"age\" \"age\", \"DATART_VTABLE\".\"id\" AS \"id\", SUM(\"DATART_VTABLE\".\"val\") \"SUM(val)\"  " +
                 "FROM  ( &#%xxx%#& )  \"DATART_VTABLE\"  GROUP BY \"DATART_VTABLE\".\"id\"  ORDER BY COUNT(\"DATART_VTABLE\".\"age\") DESC");
     }
