+++ data-providers/jdbc-data-provider/src/test/java/datart/data/provider/sql/examples/SpecialSqlExamples.java
